# apps/miniapp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Telegram Bot Manager API",
    "description": "REST API behind the Telegram Mini App admin panel.\n\nAuthenticate by posting the WebApp `initData` string to `/api/auth/telegram`,\nthen send the returned token as `Authorization: Bearer <token>` on every other\nrequest. Tokens are short-lived; renew them with `/api/auth/refresh` rather than\nre-posting `initData`, which Telegram only lets you use once.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chats/{chat_id}/triggers": {
      "get": {
        "tags": [
          "triggers"
        ],
        "summary": "Every trigger rule in a chat",
        "description": "Disabled rules included — the panel renders them with the switch off.",
        "operationId": "listTriggers",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TriggerEntry"
                  },
                  "title": "Response Listtriggers"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "402": {
            "description": "The chat's plan does not include this feature.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "triggers"
        ],
        "summary": "Add a trigger rule",
        "operationId": "createTrigger",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerEntry"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "402": {
            "description": "The chat's plan does not include this feature.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{chat_id}/triggers/{trigger_id}": {
      "get": {
        "tags": [
          "triggers"
        ],
        "summary": "One trigger rule",
        "operationId": "getTrigger",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Trigger id, unique per chat.",
              "title": "Trigger Id"
            },
            "description": "Trigger id, unique per chat."
          },
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerEntry"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "402": {
            "description": "The chat's plan does not include this feature.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "triggers"
        ],
        "summary": "Change a trigger rule",
        "operationId": "updateTrigger",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Trigger id, unique per chat.",
              "title": "Trigger Id"
            },
            "description": "Trigger id, unique per chat."
          },
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerEntry"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "402": {
            "description": "The chat's plan does not include this feature.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "triggers"
        ],
        "summary": "Delete a trigger rule",
        "description": "Not plan-gated: a downgraded chat must still be able to clear its rules.",
        "operationId": "deleteTrigger",
        "security": [
          {
            "TelegramSession": []
          }
        ],
        "parameters": [
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Trigger id, unique per chat.",
              "title": "Trigger Id"
            },
            "description": "Trigger id, unique per chat."
          },
          {
            "name": "chat_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Chat Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            }
          },
          "401": {
            "description": "Session token missing, malformed or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "402": {
            "description": "The chat's plan does not include this feature.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such chat, or the caller does not administer it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Request failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OperationResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          }
        },
        "type": "object",
        "title": "OperationResult"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "domain-error"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "RFC7807-like error body returned by the API exception handler.\n\n`title` is localized for the caller and safe to show; `detail` is the\ndeveloper-facing message. `code` is the stable machine-readable discriminator\nthe Mini App branches on, and `context` carries whatever the error knew (the\nrequired plan for a paywall, the offending field for a validation failure)."
      },
...
      "TriggerButton": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Text"
          },
          "url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "text",
          "url"
        ],
        "title": "TriggerButton"
      },
      "TriggerCreate": {
        "properties": {
          "pattern": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Pattern"
          },
          "match": {
            "$ref": "#/components/schemas/TriggerMatch",
            "default": "contains"
          },
          "response": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Response"
          },
          "media_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media File Id"
          },
          "buttons": {
            "items": {
              "$ref": "#/components/schemas/TriggerButton"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Buttons"
          },
          "case_sensitive": {
            "type": "boolean",
            "title": "Case Sensitive",
            "default": false
          },
          "delete_trigger": {
            "type": "boolean",
            "title": "Delete Trigger",
            "default": false
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "pattern",
          "response"
        ],
        "title": "TriggerCreate"
      },
      "TriggerEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "match": {
            "$ref": "#/components/schemas/TriggerMatch"
          },
          "response": {
            "type": "string",
            "title": "Response"
          },
          "media_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media File Id"
          },
          "buttons": {
            "items": {
              "$ref": "#/components/schemas/TriggerButton"
            },
            "type": "array",
            "title": "Buttons"
          },
          "case_sensitive": {
            "type": "boolean",
            "title": "Case Sensitive",
            "default": false
          },
          "delete_trigger": {
            "type": "boolean",
            "title": "Delete Trigger",
            "default": false
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "hits": {
            "type": "integer",
            "title": "Hits",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "pattern",
          "match",
          "response"
        ],
        "title": "TriggerEntry"
      },
      "TriggerMatch": {
        "type": "string",
        "enum": [
          "exact",
          "contains",
          "regex"
        ],
        "title": "TriggerMatch"
      },
      "TriggerUpdate": {
        "properties": {
          "pattern": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriggerMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "response": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "media_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media File Id"
          },
          "buttons": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TriggerButton"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buttons"
          },
          "case_sensitive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Sensitive"
          },
          "delete_trigger": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delete Trigger"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "TriggerUpdate"
      },
...
    },
    "securitySchemes": {
      "TelegramSession": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}